+++ hitsdb-20200615/alibabacloud_hitsdb20200615/client.py
@@ -2913,6 +2913,130 @@ class Client(OpenApiClient):
         runtime = util_models.RuntimeOptions()
         return await self.get_ldps_resource_cost_with_options_async(request, runtime)
 
+    def get_lindorm_engine_config_with_options(
+        self,
+        request: hitsdb_20200615_models.GetLindormEngineConfigRequest,
+        runtime: util_models.RuntimeOptions,
+    ) -> hitsdb_20200615_models.GetLindormEngineConfigResponse:
+        """
+        @summary 获取Lindorm引擎配置
+        
+        @param request: GetLindormEngineConfigRequest
+        @param runtime: runtime options for this request RuntimeOptions
+        @return: GetLindormEngineConfigResponse
+        """
+        UtilClient.validate_model(request)
+        query = {}
+        if not UtilClient.is_unset(request.engine_type):
+            query['EngineType'] = request.engine_type
+        if not UtilClient.is_unset(request.instance_id):
+            query['InstanceId'] = request.instance_id
+        if not UtilClient.is_unset(request.owner_account):
+            query['OwnerAccount'] = request.owner_account
+        if not UtilClient.is_unset(request.owner_id):
+            query['OwnerId'] = request.owner_id
+        if not UtilClient.is_unset(request.region_id):
+            query['RegionId'] = request.region_id
+        if not UtilClient.is_unset(request.resource_owner_account):
+            query['ResourceOwnerAccount'] = request.resource_owner_account
+        if not UtilClient.is_unset(request.resource_owner_id):
+            query['ResourceOwnerId'] = request.resource_owner_id
+        if not UtilClient.is_unset(request.security_token):
+            query['SecurityToken'] = request.security_token
+        req = open_api_models.OpenApiRequest(
+            query=OpenApiUtilClient.query(query)
+        )
+        params = open_api_models.Params(
+            action='GetLindormEngineConfig',
+            version='2020-06-15',
+            protocol='HTTPS',
+            pathname='/',
+            method='POST',
+            auth_type='AK',
+            style='RPC',
+            req_body_type='formData',
+            body_type='json'
+        )
+        return TeaCore.from_map(
+            hitsdb_20200615_models.GetLindormEngineConfigResponse(),
+            self.call_api(params, req, runtime)
+        )
+
+    async def get_lindorm_engine_config_with_options_async(
+        self,
+        request: hitsdb_20200615_models.GetLindormEngineConfigRequest,
+        runtime: util_models.RuntimeOptions,
+    ) -> hitsdb_20200615_models.GetLindormEngineConfigResponse:
+        """
+        @summary 获取Lindorm引擎配置
+        
+        @param request: GetLindormEngineConfigRequest
+        @param runtime: runtime options for this request RuntimeOptions
+        @return: GetLindormEngineConfigResponse
+        """
+        UtilClient.validate_model(request)
+        query = {}
+        if not UtilClient.is_unset(request.engine_type):
+            query['EngineType'] = request.engine_type
+        if not UtilClient.is_unset(request.instance_id):
+            query['InstanceId'] = request.instance_id
+        if not UtilClient.is_unset(request.owner_account):
+            query['OwnerAccount'] = request.owner_account
+        if not UtilClient.is_unset(request.owner_id):
+            query['OwnerId'] = request.owner_id
+        if not UtilClient.is_unset(request.region_id):
+            query['RegionId'] = request.region_id
+        if not UtilClient.is_unset(request.resource_owner_account):
+            query['ResourceOwnerAccount'] = request.resource_owner_account
+        if not UtilClient.is_unset(request.resource_owner_id):
+            query['ResourceOwnerId'] = request.resource_owner_id
+        if not UtilClient.is_unset(request.security_token):
+            query['SecurityToken'] = request.security_token
+        req = open_api_models.OpenApiRequest(
+            query=OpenApiUtilClient.query(query)
+        )
+        params = open_api_models.Params(
+            action='GetLindormEngineConfig',
+            version='2020-06-15',
+            protocol='HTTPS',
+            pathname='/',
+            method='POST',
+            auth_type='AK',
+            style='RPC',
+            req_body_type='formData',
+            body_type='json'
+        )
+        return TeaCore.from_map(
+            hitsdb_20200615_models.GetLindormEngineConfigResponse(),
+            await self.call_api_async(params, req, runtime)
+        )
+
+    def get_lindorm_engine_config(
+        self,
+        request: hitsdb_20200615_models.GetLindormEngineConfigRequest,
+    ) -> hitsdb_20200615_models.GetLindormEngineConfigResponse:
+        """
+        @summary 获取Lindorm引擎配置
+        
+        @param request: GetLindormEngineConfigRequest
+        @return: GetLindormEngineConfigResponse
+        """
+        runtime = util_models.RuntimeOptions()
+        return self.get_lindorm_engine_config_with_options(request, runtime)
+
+    async def get_lindorm_engine_config_async(
+        self,
+        request: hitsdb_20200615_models.GetLindormEngineConfigRequest,
+    ) -> hitsdb_20200615_models.GetLindormEngineConfigResponse:
+        """
+        @summary 获取Lindorm引擎配置
+        
+        @param request: GetLindormEngineConfigRequest
+        @return: GetLindormEngineConfigResponse
+        """
+        runtime = util_models.RuntimeOptions()
+        return await self.get_lindorm_engine_config_with_options_async(request, runtime)
+
     def get_lindorm_fs_used_detail_with_options(
         self,
         request: hitsdb_20200615_models.GetLindormFsUsedDetailRequest,
@@ -4444,6 +4568,138 @@ class Client(OpenApiClient):
         """
         runtime = util_models.RuntimeOptions()
         return await self.list_tag_resources_with_options_async(request, runtime)
+
+    def migrate_single_zone_to_multi_zone_with_options(
+        self,
+        request: hitsdb_20200615_models.MigrateSingleZoneToMultiZoneRequest,
+        runtime: util_models.RuntimeOptions,
+    ) -> hitsdb_20200615_models.MigrateSingleZoneToMultiZoneResponse:
+        """
+        @summary 单可用区迁移多可用区基础版
+        
+        @param request: MigrateSingleZoneToMultiZoneRequest
+        @param runtime: runtime options for this request RuntimeOptions
+        @return: MigrateSingleZoneToMultiZoneResponse
+        """
+        UtilClient.validate_model(request)
+        query = {}
+        if not UtilClient.is_unset(request.arbitrary_vswitch_id):
+            query['ArbitraryVSwitchId'] = request.arbitrary_vswitch_id
+        if not UtilClient.is_unset(request.arbitrary_zone_id):
+            query['ArbitraryZoneId'] = request.arbitrary_zone_id
+        if not UtilClient.is_unset(request.instance_id):
+            query['InstanceId'] = request.instance_id
+        if not UtilClient.is_unset(request.owner_account):
+            query['OwnerAccount'] = request.owner_account
+        if not UtilClient.is_unset(request.owner_id):
+            query['OwnerId'] = request.owner_id
+        if not UtilClient.is_unset(request.resource_owner_account):
+            query['ResourceOwnerAccount'] = request.resource_owner_account
+        if not UtilClient.is_unset(request.resource_owner_id):
+            query['ResourceOwnerId'] = request.resource_owner_id
+        if not UtilClient.is_unset(request.security_token):
+            query['SecurityToken'] = request.security_token
+        if not UtilClient.is_unset(request.standby_vswitch_id):
+            query['StandbyVSwitchId'] = request.standby_vswitch_id
+        if not UtilClient.is_unset(request.standby_zone_id):
+            query['StandbyZoneId'] = request.standby_zone_id
+        req = open_api_models.OpenApiRequest(
+            query=OpenApiUtilClient.query(query)
+        )
+        params = open_api_models.Params(
+            action='MigrateSingleZoneToMultiZone',
+            version='2020-06-15',
+            protocol='HTTPS',
+            pathname='/',
+            method='POST',
+            auth_type='AK',
+            style='RPC',
+            req_body_type='formData',
+            body_type='json'
+        )
+        return TeaCore.from_map(
+            hitsdb_20200615_models.MigrateSingleZoneToMultiZoneResponse(),
+            self.call_api(params, req, runtime)
+        )
+
+    async def migrate_single_zone_to_multi_zone_with_options_async(
+        self,
+        request: hitsdb_20200615_models.MigrateSingleZoneToMultiZoneRequest,
+        runtime: util_models.RuntimeOptions,
+    ) -> hitsdb_20200615_models.MigrateSingleZoneToMultiZoneResponse:
+        """
+        @summary 单可用区迁移多可用区基础版
+        
+        @param request: MigrateSingleZoneToMultiZoneRequest
+        @param runtime: runtime options for this request RuntimeOptions
+        @return: MigrateSingleZoneToMultiZoneResponse
+        """
+        UtilClient.validate_model(request)
+        query = {}
+        if not UtilClient.is_unset(request.arbitrary_vswitch_id):
+            query['ArbitraryVSwitchId'] = request.arbitrary_vswitch_id
+        if not UtilClient.is_unset(request.arbitrary_zone_id):
+            query['ArbitraryZoneId'] = request.arbitrary_zone_id
+        if not UtilClient.is_unset(request.instance_id):
+            query['InstanceId'] = request.instance_id
+        if not UtilClient.is_unset(request.owner_account):
+            query['OwnerAccount'] = request.owner_account
+        if not UtilClient.is_unset(request.owner_id):
+            query['OwnerId'] = request.owner_id
+        if not UtilClient.is_unset(request.resource_owner_account):
+            query['ResourceOwnerAccount'] = request.resource_owner_account
+        if not UtilClient.is_unset(request.resource_owner_id):
+            query['ResourceOwnerId'] = request.resource_owner_id
+        if not UtilClient.is_unset(request.security_token):
+            query['SecurityToken'] = request.security_token
+        if not UtilClient.is_unset(request.standby_vswitch_id):
+            query['StandbyVSwitchId'] = request.standby_vswitch_id
+        if not UtilClient.is_unset(request.standby_zone_id):
+            query['StandbyZoneId'] = request.standby_zone_id
+        req = open_api_models.OpenApiRequest(
+            query=OpenApiUtilClient.query(query)
+        )
+        params = open_api_models.Params(
+            action='MigrateSingleZoneToMultiZone',
+            version='2020-06-15',
+            protocol='HTTPS',
+            pathname='/',
+            method='POST',
+            auth_type='AK',
+            style='RPC',
+            req_body_type='formData',
+            body_type='json'
+        )
+        return TeaCore.from_map(
+            hitsdb_20200615_models.MigrateSingleZoneToMultiZoneResponse(),
+            await self.call_api_async(params, req, runtime)
+        )
+
+    def migrate_single_zone_to_multi_zone(
+        self,
+        request: hitsdb_20200615_models.MigrateSingleZoneToMultiZoneRequest,
+    ) -> hitsdb_20200615_models.MigrateSingleZoneToMultiZoneResponse:
+        """
+        @summary 单可用区迁移多可用区基础版
+        
+        @param request: MigrateSingleZoneToMultiZoneRequest
+        @return: MigrateSingleZoneToMultiZoneResponse
+        """
+        runtime = util_models.RuntimeOptions()
+        return self.migrate_single_zone_to_multi_zone_with_options(request, runtime)
+
+    async def migrate_single_zone_to_multi_zone_async(
+        self,
+        request: hitsdb_20200615_models.MigrateSingleZoneToMultiZoneRequest,
+    ) -> hitsdb_20200615_models.MigrateSingleZoneToMultiZoneResponse:
+        """
+        @summary 单可用区迁移多可用区基础版
+        
+        @param request: MigrateSingleZoneToMultiZoneRequest
+        @return: MigrateSingleZoneToMultiZoneResponse
+        """
+        runtime = util_models.RuntimeOptions()
+        return await self.migrate_single_zone_to_multi_zone_with_options_async(request, runtime)
 
     def modify_auto_scaling_config_with_options(
         self,